+++ src/constants/index.js
@@ -17,7 +17,6 @@ const DEFAULTS = {
   maxHeight: 250,
   position: 'bottom',
 
-  displayHtml: false,
   displayValues: false,
   displayTitle: false,
   displayDelimiter: ', ',